+++ bootstrap.py
@@ -1,9 +1,10 @@
 """
-:Version: 1.3
+:Version: 1.3.1
 
 Look for the latest version of the script in the GIST:
 https://gist.github.com/Cykooz/118fba100c9ceb76ee822f7541d480c4
 """
+
 import argparse
 import logging
 import os
@@ -18,12 +19,12 @@ from urllib.request import urlopen
 
 
 BUILDOUT_VERSION = '3.0.1'
-PIP_VERSION = '23.2.1'
-# With newer version a build is failed due to
-# incorrect version of dependencies in some packages.
+PIP_VERSION = '24.0.0'
+# With newer version, a build is failed due to
+# an incorrect version of dependencies in some packages.
 # Need to recheck in the future.
 SETUPTOOLS_VERSION = '66.1.1'
-WHEEL_VERSION = '0.41.2'
+WHEEL_VERSION = '0.43.0'
 
 GET_PIP_URL = 'https://bootstrap.pypa.io/get-pip.py'
 
@@ -31,31 +32,35 @@ GET_PIP_URL = 'https://bootstrap.pypa.io/get-pip.py'
 def main():
     parser = argparse.ArgumentParser(description='Bootstrap zc.buildout')
     parser.add_argument(
-        '-r, --recreate', dest='recreate', action='store_true',
+        '-r, --recreate',
+        dest='recreate',
+        action='store_true',
         help='Recreate virtual python if it exists.',
     )
     parser.add_argument(
-        '--buildout_version', default=BUILDOUT_VERSION,
+        '--buildout_version',
+        default=BUILDOUT_VERSION,
         help='Version of zc.buildout (default: %(default)).',
     )
     parser.add_argument(
-        '--pip_version', default=PIP_VERSION,
+        '--pip_version',
+        default=PIP_VERSION,
         help='Version of pip (default: %(default)).',
     )
     parser.add_argument(
-        '--setuptools_version', default=SETUPTOOLS_VERSION,
+        '--setuptools_version',
+        default=SETUPTOOLS_VERSION,
         help='Version of setuptools (default: %(default)).',
     )
     parser.add_argument(
-        '--wheel_version', default=WHEEL_VERSION,
+        '--wheel_version',
+        default=WHEEL_VERSION,
         help='Version of wheel (default: %(default)).',
     )
 
     args = parser.parse_args(sys.argv[1:])
     logging.basicConfig(
-        stream=sys.stdout,
-        level=logging.DEBUG,
-        format='%(levelname)s: %(message)s'
+        stream=sys.stdout, level=logging.DEBUG, format='%(levelname)s: %(message)s'
     )
     logger = logging.getLogger('buildout.bootstrap')
 
@@ -79,7 +84,7 @@ def main():
                 '-c',
                 'import sys;print(sys.version)',
             ],
-            stdout=subprocess.PIPE
+            stdout=subprocess.PIPE,
         )
         stdout, stderr = proc.communicate()
         env_version = stdout.strip().decode()
@@ -134,11 +139,15 @@ def main():
     dependencies = [p + v for p, v in dependencies.items()]
     if dependencies:
         logger.info('Installing ' + ', '.join(dependencies))
-        _run_cmd([
-                     str(venv_python_path),
-                     '-m', 'pip',
-                     'install',
-                 ] + dependencies)
+        _run_cmd(
+            [
+                str(venv_python_path),
+                '-m',
+                'pip',
+                'install',
+            ]
+            + dependencies
+        )
 
     logger.info('Bootstrap zc.buildout')
     is_win = platform.system() == 'Windows'
@@ -146,10 +155,12 @@ def main():
         buildout_path = venv_dir / 'Scripts' / 'buildout.exe'
     else:
         buildout_path = venv_dir / 'bin' / 'buildout'
-    _run_cmd([
-        str(buildout_path),
-        'bootstrap',
-    ])
+    _run_cmd(
+        [
+            str(buildout_path),
+            'bootstrap',
+        ]
+    )
 
     # Delete zc.buildout.egg-link from list of develop eggs
     buildout_link = project_dir / 'develop-eggs' / 'zc.buildout.egg-link'